Have proper message displayed upon failure

+++ modules/data_release/ajax/FileUpload.php
@@ -31,27 +31,35 @@ if ($_GET['action'] == 'upload') {
 
     $baseURL = $settings->getBaseURL();
     $path    = \Utility::appendForwardSlash($config->getSetting('dataReleasePath'));
-
-    $target_path = $path . $fileName;
-    if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_path)) {
-        // insert the file into the data_release table
-        $DB->insert(
-            'data_release',
-            array(
-                'file_name'   => $fileName,
-                'version'     => $version,
-                'upload_date' => $upload_date,
-            )
-        );
-        // get the ID of the user who uploaded the file
-        $user_ID = $DB->pselectOne(
-            "SELECT ID FROM users WHERE userid=:UserID",
-            array('UserID' => $user->getUsername())
-        );
-        // get the ID of the file inserted in the data_release table
-        $version_where = $version ? "version=:version" : "version IS :version";
-        $ID            = $DB->pselectOne(
-            "SELECT 
+    if (!file_exists($path)) {
+        $msg = "File upload failed. Upload directory not found.";
+        error_log('ERROR: ' . $msg);
+        header("HTTP/1.1 404 $msg");
+    } elseif (!is_writable($path)) {
+        $msg = "File upload failed. Upload directory is not writeable.";
+        error_log('ERROR: ' . $msg);
+        header("HTTP/1.1 404 $msg");
+    } else {
+        $target_path = $path . $fileName;
+        if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_path)) {
+            // insert the file into the data_release table
+            $DB->insert(
+                'data_release',
+                array(
+                    'file_name' => $fileName,
+                    'version' => $version,
+                    'upload_date' => $upload_date,
+                )
+            );
+            // get the ID of the user who uploaded the file
+            $user_ID = $DB->pselectOne(
+                "SELECT ID FROM users WHERE userid=:UserID",
+                array('UserID' => $user->getUsername())
+            );
+            // get the ID of the file inserted in the data_release table
+            $version_where = $version ? "version=:version" : "version IS :version";
+            $ID = $DB->pselectOne(
+                "SELECT 
                id 
              FROM 
                data_release 
@@ -59,23 +67,24 @@ if ($_GET['action'] == 'upload') {
                file_name=:file_name 
                AND $version_where
                AND upload_date=:upload_date",
-            array(
-                'file_name'   => $fileName,
-                'version'     => $version,
-                'upload_date' => $upload_date,
-            )
-        );
-        // add permission to the user for the uploaded data_release file
-        $DB->insert(
-            'data_release_permissions',
-            array(
-                'userid'          => $user_ID,
-                'data_release_id' => $ID,
-            )
-        );
+                array(
+                    'file_name' => $fileName,
+                    'version' => $version,
+                    'upload_date' => $upload_date,
+                )
+            );
+            // add permission to the user for the uploaded data_release file
+            $DB->insert(
+                'data_release_permissions',
+                array(
+                    'userid' => $user_ID,
+                    'data_release_id' => $ID,
+                )
+            );
+        }
+        header("Location: {$baseURL}/data_release/?uploadSuccess=true");
+        header("HTTP/1.1 201 Created");
     }
-    header("Location: {$baseURL}/data_release/?uploadSuccess=true");
-    header("HTTP/1.1 201 Created");
 
 } elseif ($_GET['action'] == 'getData') {
     $filesList = $DB->pselect(
